Abstracted everything for DashBoard to it's own file. Did some minor cleanup and naming.
